Telegram_Bot_Translator/requirements.txt
python-telegram-bot==21.6
deep-translator==1.11.4
python-dotenv==1.0.1